restructure alert-silence command
