+++ cmd/alerts/silence/add_silence.go
@@ -3,47 +3,125 @@ package silence
 import (
 	"fmt"
 	"log"
+	"strings"
 
+	ocmutils "github.com/openshift/osdctl/pkg/utils"
 	"github.com/spf13/cobra"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
 )
 
-func NewCmdAddSilence() *cobra.Command {
-	return &cobra.Command{
-		Use:   "add <cluster-id> <alertname1> <alertname2>... <comment>",
-		Short: "add a new silence",
-		Long:  `add new silence`,
-		Args:  cobra.MinimumNArgs(3),
-		Run: func(cmd *cobra.Command, args []string) {
-			clusterID := args[0]
-			comment := args[len(args)-1]
-			alertNames := args[1 : len(args)-1]
-			AddSilence(clusterID, alertNames, comment)
-		},
-	}
+type addSilenceCmd struct {
+	clusterID string
+	alertID   []string
+	duration  string
+	comment   string
+	all       bool
 }
 
-// osdctl alerts add-silence ${CLUSTERID} ${alertname1} ${alertname2}... ${comment}
-func AddSilence(clusterID string, alertNames []string, comment string) {
+func NewCmdAddSilence() *cobra.Command {
+	addSilenceCmd := &addSilenceCmd{}
+	cmd := &cobra.Command{
+		Use:               "add <cluster-id> [--all --duration --comment | --alertname --duration --comment]",
+		Short:             "Add new silence for alert",
+		Long:              `add new silence for specfic or all alert with comment and duration of alert`,
+		Args:              cobra.ExactArgs(1),
+		DisableAutoGenTag: true,
+		Run: func(cmd *cobra.Command, args []string) {
+			addSilenceCmd.clusterID = args[0]
+			AddSilence(addSilenceCmd)
+		},
+	}
+
+	cmd.Flags().StringSliceVar(&addSilenceCmd.alertID, "alertname", []string{}, "alertname (comma-separated)")
+	cmd.Flags().StringVarP(&addSilenceCmd.comment, "comment", "c", "", "add comment about silence")
+	cmd.Flags().StringVarP(&addSilenceCmd.duration, "duration", "d", "15d", "add duration for silence")
+	//default duration set to 15 days
+	cmd.Flags().BoolVarP(&addSilenceCmd.all, "all", "a", false, "add silences for all alert")
+
+	return cmd
+}
+
+func AddSilence(cmd *addSilenceCmd) {
+	username := GetUserInfo()
+	clusterID := cmd.clusterID
+	alertID := cmd.alertID
+	comment := cmd.comment
+	duration := cmd.duration
+	all := cmd.all
+
 	kubeconfig, clientset, err := GetKubeConfigClient(clusterID)
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	for _, alertname := range alertNames {
+	if all {
+		AddAllSilence(clusterID, duration, comment, username, kubeconfig, clientset)
+	} else if len(alertID) > 0 {
+		AddAlertNameSilence(alertID, duration, comment, username, kubeconfig, clientset)
+	} else {
+		fmt.Println("No valid option specified. Use --all or --alertname.")
+	}
+
+}
+
+func AddAllSilence(clusterID, duration, comment, username string, kubeconfig *rest.Config, clientset *kubernetes.Clientset) {
+	addCmd := []string{
+		"amtool",
+		"silence",
+		"add",
+		"cluster=" + clusterID,
+		"--alertmanager.url=" + LocalHostUrl,
+		"--duration=" + duration,
+		"--comment=" + comment,
+	}
+
+	output, err := ExecInPod(kubeconfig, clientset, addCmd)
+	if err != nil {
+		fmt.Println(err)
+	}
+
+	formattedOutput := strings.Replace(output, "\n", " ", -1)
+
+	fmt.Printf("All alerts for cluster %s has been silenced with id \"%s\"for a duration of %s by user %s.\n", ClusterName, formattedOutput, duration, username)
+}
+
+func AddAlertNameSilence(alertID []string, duration, comment, username string, kubeconfig *rest.Config, clientset *kubernetes.Clientset) {
+	for _, alertname := range alertID {
 		addCmd := []string{
 			"amtool",
 			"silence",
 			"add",
 			"alertname=" + alertname,
 			"--alertmanager.url=" + LocalHostUrl,
+			"--duration=" + duration,
 			"--comment=" + comment,
 		}
 
-		output, err := ExecInPod(kubeconfig, clientset, LocalHostUrl, addCmd, PodName)
+		output, err := ExecInPod(kubeconfig, clientset, addCmd)
 		if err != nil {
 			fmt.Println(err)
 		}
 
-		fmt.Printf("%v", output)
+		formattedOutput := strings.Replace(output, "\n", " ", -1)
+
+		fmt.Printf("Alert \"%s\" has been silenced with id \"%s\"for duration of %s by user %s.\n", alertname, formattedOutput, duration, username)
 	}
+}
+
+// User name of who added the silence
+func GetUserInfo() string {
+	connection, err := ocmutils.CreateConnection()
+	if err != nil {
+		fmt.Printf("Error %s in create connection.", err)
+	}
+	defer connection.Close()
+
+	account, err := connection.AccountsMgmt().V1().CurrentAccount().Get().Send()
+	if err != nil {
+		fmt.Printf("Error %s in retreving the details of account.", err)
+	}
+
+	name, _ := account.Body().GetUsername()
+	return name
 }
