+++ cmd/alerts/silence/clear_silence.go
@@ -3,8 +3,11 @@ package silence
 import (
 	"fmt"
 	"log"
+	"strings"
 
 	"github.com/spf13/cobra"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
 )
 
 type silenceCmd struct {
@@ -16,28 +19,22 @@ type silenceCmd struct {
 func NewCmdClearSilence() *cobra.Command {
 	silenceCmd := &silenceCmd{}
 	cmd := &cobra.Command{
-		Use:               "clear <cluster-id> [--all | --silenceID <silence-id>]",
-		Short:             "clear all silence",
-		Long:              `clear all created silence`,
+		Use:               "expire <cluster-id> [--all | --silence-id <silence-id>]",
+		Short:             "Expire Silence for alert",
+		Long:              `expire all silence or based on silenceid`,
 		Args:              cobra.ExactArgs(1),
 		DisableAutoGenTag: true,
 		Run: func(cmd *cobra.Command, args []string) {
-			//silenceCmd.clusterID = args[0]
-			//silenceID := args[1 : len(args)-1]
-			//ClearSilence(silenceCmd.clusterID, silenceCmd.silenceID)
 			silenceCmd.clusterID = args[0]
 			ClearSilence(silenceCmd)
 		},
 	}
-	cmd.Flags().StringSliceVar(&silenceCmd.silenceID, "silenceID", []string{}, "silence id (comma-separated)")
-	//cmd.Flags().BoolP("all", "a", false, "clear all silences")
-	cmd.Flags().BoolVar(&silenceCmd.all, "all", false, "clear all silences")
+	cmd.Flags().StringSliceVar(&silenceCmd.silenceID, "silence-id", []string{}, "silence id (comma-separated)")
+	cmd.Flags().BoolVarP(&silenceCmd.all, "all", "a", false, "clear all silences")
 	return cmd
 }
 
-// osdctl alerts clear-silence ${CLUSTERID} --silenceID/--all
 func ClearSilence(cmd *silenceCmd) {
-	var newID string 
 	clusterID := cmd.clusterID
 	silenceID := cmd.silenceID
 	all := cmd.all
@@ -46,46 +43,87 @@ func ClearSilence(cmd *silenceCmd) {
 	if err != nil {
 		log.Fatal(err)
 	}
-	
-	for _, silenceId := range silenceID {
-		if all {
-			clearCmd := []string{
-				"amtool",
-				"silence",
-				"expire",
-				newID,
-				"--alertmanager.url=" + LocalHostUrl,
-			}
 
-			output, err := ExecInPod(kubeconfig, clientset, LocalHostUrl, clearCmd, PodName)
-
-			if err != nil {
-				log.Printf("Error expiring all silences: %v\n", err)
-				continue
-			}
-
-			fmt.Println("All silences expired successfully.")
-			fmt.Printf("amtool output: %v\n", output)
-
+	if all {
+		ClearAllSilence(kubeconfig, clientset)
+	} else {
+		if len(silenceID) > 0 {
+			ClearSilenceByID(silenceID, kubeconfig, clientset)
 		} else {
-			// Assuming SilenceID is a string, replace with the actual field from your Silence struct
-			clearCmd := []string{
-				"amtool",
-				"silence",
-				"expire",
-				silenceId,
-				"--alertmanager.url=" + LocalHostUrl,
-			}
-
-			output, err := ExecInPod(kubeconfig, clientset, LocalHostUrl, clearCmd, PodName)
-
-			if err != nil {
-				log.Printf("Error expiring silence ID %s: %v\n", silenceId, err)
-				continue
-			}
-
-			fmt.Printf("Silence ID %s expired successfully.\n", silenceId)
-			fmt.Printf("amtool output: %v\n", output)
+			fmt.Println("No valid option specified. Using a default option")
+			ClearAllSilence(kubeconfig, clientset)
 		}
+	}
+}
+
+func ClearAllSilence(kubeconfig *rest.Config, clientset *kubernetes.Clientset) {
+	queryCmd := []string{
+		"amtool",
+		"silence",
+		"query",
+		"-q",
+		"--alertmanager.url=" + LocalHostUrl,
+	}
+
+	queryOutput, err := ExecInPod(kubeconfig, clientset, queryCmd)
+
+	if err != nil {
+		fmt.Print("some issue in query command")
+	}
+
+	formattedOutput := strings.Join(strings.Fields(queryOutput), " ")
+
+	if formattedOutput == " " || formattedOutput == "" {
+		fmt.Println("No Silence has been set for alerts, please create new silence")
+		return
+	}
+
+	finalOutput := strings.Fields(formattedOutput)
+	countsilence := len(finalOutput)
+	for _, silence := range finalOutput {
+		clearCmd := []string{
+			"amtool",
+			"silence",
+			"expire",
+			silence,
+			"--alertmanager.url=" + LocalHostUrl,
+		}
+
+		countsilence = countsilence - 1
+
+		_, err := ExecInPod(kubeconfig, clientset, clearCmd)
+
+		if err != nil {
+			log.Printf("Error expiring silence ID \"%s\" : %v\n", silence, err)
+			return
+		}
+
+		fmt.Printf("SilenceID \"%s\" expired successfully.\n", silence)
+
+		if countsilence == 0 {
+			fmt.Println()
+			fmt.Printf("All SilenceID expired successfully.\n")
+			return
+		}
+	}
+}
+
+func ClearSilenceByID(silenceID []string, kubeconfig *rest.Config, clientset *kubernetes.Clientset) {
+	for _, silenceId := range silenceID {
+		clearCmd := []string{
+			"amtool",
+			"silence",
+			"expire",
+			silenceId,
+			"--alertmanager.url=" + LocalHostUrl,
+		}
+		_, err := ExecInPod(kubeconfig, clientset, clearCmd)
+
+		if err != nil {
+			log.Printf("Error expiring silence ID \"%s\" %v\n", silenceId, err)
+			continue
+		}
+
+		fmt.Printf("Requested SilenceID \"%s\" expired successfully.\n", silenceId)
 	}
 }
